requirements.txt
pandas>=1.5.3
numpy>=1.23.5
scikit-learn>=1.2.2
paho-mqtt>=1.6.1
joblib>=1.2.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.1
websockets>=11.0.3
setuptools>=65.5.0
wheel>=0.38.4